*** empty log message ***
git-svn-id: svn://localhost/trunk@955 42a48d6c-1186-114d-bcb1-a7d39102e91d

+++ com.python.pydev/src/edu/umd/cs/piccolox/swt/SWTGraphics2D.java
@@ -238,7 +238,7 @@ public class SWTGraphics2D extends Graphics2D {
 					style = style | SWT.ITALIC;	
 				}
 				
-				return new Font(fd[0].getName(),style,fd[0].height);
+				return new Font(fd[0].getName(),style,(int)fd[0].height);
 			}
 			return null;			
 		}
